Always perform NAT in NAT-supported hooks
This avoids a race condition when a packet is being written and the NAT
table is being updated at the same time.

Previously, NAT will only be skipped if either the connection has been
finalized or the hook's relevant NAT (DNAT for Prerouting/Output; SNAT
for Input/Postrouting) has been performed. However, it is possible for
the following sequence of events to occur:

  1) A packet performs DNAT related hooks in Prerouting or Output
     but not perform DNAT as no rule matched the packet.
  2) The NAT table updates such that a DNAT rule now will be performed
     on packets matching the packet's tuple from (1).
  3) A second packet matching the original packet's tuple performs
     the Prerouting or Output hook, now having performed DNAT and
     updating the connection.
  4) Either packet goes through the other hooks and finalizes the
     connection.

Here we would have 2 packets that have the same original tuple but have
different destination address/ports after performing all the hooks.
Later packets will look like the second packet in the example but the
first packet may trigger a response that the connection table will not
recognize, potentially leading to an ICMP error or TCP RST.

A similar race exists for SNAT.

To avoid the race, this change guarantees that {D,S}NAT is always
performed on a connection before leaving the relevant hook. This
way we make sure that all packets that are associated with a
connection will have the same tuple, per direction.

PiperOrigin-RevId: 409320740